+++ Vault-Coder/Changes.md
@@ -8,6 +8,7 @@
 - Added support for targeting access -writing
   - This feature is enabled by passing `-T` as a console argument
 - Added **UncertainBoolean** data type
+- Added **Year** and **Month** data types
 ### Bugfixes
 - Fixed a bug that would cause an infinite loop when two classes directly reference each other
 - Fixed a bug that would cause certain inheritance-based combo classes to interrupt the code-generation process